+++ js/constantData.js
@@ -1,0 +1,54 @@
+const EMPLOYES_COUNT = 25;
+const COUNT_AVATAR = 6;
+const MAXIMUM_MESSAGES = 2;
+const MAX_LIKES = 200;
+const MIN_LIKES = 15;
+const MAX_COUNT_UNIK_ID = 10000;
+const MAX_COUNT_COMMENTS = 10;
+const DESCRIPTION_INTRODUCION = [
+  'На фотографии мы видим…',
+  'При первом взгляде на фотографию становится очевидным, что…',
+  'Предложенная для описания фотография интересна тем, что...',
+  'Если внимательно посмотреть на снимок, то…'];
+const DESCRIPTION_BASE = [
+  'На переднем плане изображены…',
+  'Если внимательно посмотреть на изображение, то…',
+  'Композиция снимка весьма интересна, потому что…',
+  'Выражения лиц у героев фотографии говорят о том, что…',
+  'Фото несёт в себе определённое настроение:…'];
+const DESCRIPTION_END = [
+  'Завершая описание, хочется отметить…',
+  'Своё описание хочется закончить…',
+  'ельзя не увидеть талант фотографа, который сумел…',
+];
+const MESSAGES_ARRAY = [
+  'Всё отлично!',
+  'В целом всё неплохо.',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+const NAMES_ARRAY = [
+  'Ника', 'Таисия', 'Михаил', 'Евгения', 'Василиса', 'Мария', 'Алиса', 'Илья', 'Иван',
+  'Милана', 'Надежда', 'Ксения', 'Амина', 'Николай', 'Кирилл', 'Алексей', 'Руслан',
+  'Олег', 'Степан', 'Ульяна', 'Мирослава', 'Андрей', 'Василий', 'Марк', 'София'
+];
+
+export {
+  EMPLOYES_COUNT,
+  COUNT_AVATAR,
+  MAXIMUM_MESSAGES,
+  MAX_LIKES,
+  MIN_LIKES,
+  MAX_COUNT_UNIK_ID,
+  MAX_COUNT_COMMENTS,
+  DESCRIPTION_INTRODUCION,
+  DESCRIPTION_BASE,
+  DESCRIPTION_END,
+  MESSAGES_ARRAY,
+  NAMES_ARRAY
+};
